+++ docs/en/6.13.0/html/group___pde___p_d_fix___s_d_k.js
@@ -1,0 +1,151 @@
+var group___pde___p_d_fix___s_d_k =
+[
+    [ "PdsTextRun", "struct_pds_text_run.html", [
+      [ "GetFirstCharIndex", "struct_pds_text_run.html#a893abc3a1baaf49a4d571369dc6ff3f1", null ],
+      [ "GetLastCharIndex", "struct_pds_text_run.html#a492fa40f51ebddc2e4465b9e2fe7b06a", null ],
+      [ "GetTextObject", "struct_pds_text_run.html#a11cd36c4ea777ea88fdafb03535f9ef3", null ]
+    ] ],
+    [ "PdsWord", "struct_pds_word.html", [
+      [ "GetBBox", "struct_pds_word.html#a27b961e8e93fb15f0768b09c448107d7", null ],
+      [ "GetCharBBox", "struct_pds_word.html#af239ac542a702478a9e9ed2b8d6850d9", null ],
+      [ "GetCharCode", "struct_pds_word.html#a82be631ff4b2a14703419a4d6aa148e0", null ],
+      [ "GetCharText", "struct_pds_word.html#adc012ba5a5a15a786927d95e634ccb07", null ],
+      [ "GetCharTextState", "struct_pds_word.html#a6a180d3ef5e2bf1ac70766be9de6ea88", null ],
+      [ "GetNumChars", "struct_pds_word.html#a403a347af334627ab4405e98deab449b", null ],
+      [ "GetNumTextRuns", "struct_pds_word.html#afd9033e8993048cee266a160c8fc68f9", null ],
+      [ "GetQuad", "struct_pds_word.html#a921ac097df6f1fd79262b60148e470fe", null ],
+      [ "GetText", "struct_pds_word.html#a9101cbb38770be1b44258265f52b1230", null ],
+      [ "GetTextRunAt", "struct_pds_word.html#a6da89f0544684375f26e33e927a658d5", null ]
+    ] ],
+    [ "PdsWordList", "struct_pds_word_list.html", [
+      [ "GetNumWords", "struct_pds_word_list.html#acfaf8f6f1ad4f62c76c35a5b04900a9a", null ],
+      [ "GetWord", "struct_pds_word_list.html#a5d40bb148e920acb8dac637651ae1319", null ],
+      [ "Release", "struct_pds_word_list.html#ac2210828044905dedf70d759ff3b7b56", null ]
+    ] ],
+    [ "PdeElement", "struct_pde_element.html", [
+      [ "GetAlignment", "struct_pde_element.html#a4b83c365df34b54bf2715b6570f31e77", null ],
+      [ "GetAngle", "struct_pde_element.html#a7d30e09b17252d7d3aec7e3ec2c7822f", null ],
+      [ "GetBBox", "struct_pde_element.html#a065871d0f4bd3d01e54794b5e14d0b2b", null ],
+      [ "GetChild", "struct_pde_element.html#ab42a29f360d8e1b205b9c3e2584c74e8", null ],
+      [ "GetData", "struct_pde_element.html#a633819805ce71919512f98b287937d7c", null ],
+      [ "GetFlags", "struct_pde_element.html#a1043ad4e6af0a323e519329ac0c4c549", null ],
+      [ "GetGraphicState", "struct_pde_element.html#a553507c3db7ae09a24029c71a0a78059", null ],
+      [ "GetId", "struct_pde_element.html#a33d921be337f11d47ca260d5b7577c54", null ],
+      [ "GetNumChildren", "struct_pde_element.html#ab6069e45929da5c16e625003b16435c6", null ],
+      [ "GetNumPageObjects", "struct_pde_element.html#a53d3de3bc85aec96a66bcfcdbecec17e", null ],
+      [ "GetPageMap", "struct_pde_element.html#a1220b2f05113d313eafb99d5d4c99d1c", null ],
+      [ "GetPageObject", "struct_pde_element.html#adbceecb9079edcbd9824aea5de27538d", null ],
+      [ "GetQuad", "struct_pde_element.html#a332a8ddd55113ab535d6146d0339d56a", null ],
+      [ "GetStateFlags", "struct_pde_element.html#a49e0bdf15fc4dba551c66324985fa3e8", null ],
+      [ "GetType", "struct_pde_element.html#a3b848c0560d9603ead89f883ddd2703b", null ],
+      [ "SetActualText", "struct_pde_element.html#a9bb63f989b3daf0ed2ccd249205a666f", null ],
+      [ "SetAlt", "struct_pde_element.html#a60e71e645253156907d55c397ae79b86", null ],
+      [ "SetBBox", "struct_pde_element.html#af3ad42234f6186c334c2a20abe62123e", null ],
+      [ "SetData", "struct_pde_element.html#aa80ae9e64f6dc9592dcf4a32bb302b1d", null ],
+      [ "SetFlags", "struct_pde_element.html#aacbaeec8b367178bb9eb1e4e101b0561", null ],
+      [ "SetStateFlags", "struct_pde_element.html#a4665f1beb5e4748d017326d7c31146a5", null ]
+    ] ],
+    [ "PdeContainer", "struct_pde_container.html", null ],
+    [ "PdeList", "struct_pde_list.html", null ],
+    [ "PdeToc", "struct_pde_toc.html", null ],
+    [ "PdeAnnot", "struct_pde_annot.html", [
+      [ "GetAnnot", "struct_pde_annot.html#ae16e696d6c8b003d0eacbc8bf23f48b6", null ]
+    ] ],
+    [ "PdeFormField", "struct_pde_form_field.html", null ],
+    [ "PdeImage", "struct_pde_image.html", [
+      [ "GetCaption", "struct_pde_image.html#a35b1b119b5bb33ff575f863526160e67", null ],
+      [ "GetImageType", "struct_pde_image.html#a9a4855616d801554afc131b792fc5282", null ]
+    ] ],
+    [ "PdeLine", "struct_pde_line.html", null ],
+    [ "PdeRect", "struct_pde_rect.html", null ],
+    [ "PdeHeader", "struct_pde_header.html", null ],
+    [ "PdeFooter", "struct_pde_footer.html", null ],
+    [ "PdeCell", "struct_pde_cell.html", [
+      [ "GetColSpan", "struct_pde_cell.html#ac5636b26baa5cb3122c7c95513330159", null ],
+      [ "GetRowSpan", "struct_pde_cell.html#a562a96210ded4bf7ce07618de21a1954", null ],
+      [ "GetSpanCell", "struct_pde_cell.html#a969baa9ad608c9534cc6471000c2e6ee", null ],
+      [ "HasBorderGraphicState", "struct_pde_cell.html#a7d28f917e0832b5ad26995830e57fc69", null ]
+    ] ],
+    [ "PdeTable", "struct_pde_table.html", [
+      [ "GetCaption", "struct_pde_table.html#a08bc6219fa125b38ddf425f3bd375112", null ],
+      [ "GetCell", "struct_pde_table.html#a5af6184776ad7d4acc84f420b71f66c1", null ],
+      [ "GetColAlignment", "struct_pde_table.html#a93f937a0460468b6492f1f107ffe61cf", null ],
+      [ "GetNumCols", "struct_pde_table.html#acdffdfd6c9e9b45a6e10c8adaba7db0b", null ],
+      [ "GetNumRows", "struct_pde_table.html#ac2adbfa23aa282831ba5f03b89893255", null ],
+      [ "GetRowAlignment", "struct_pde_table.html#aa1732a4d6c0d8f42d43dd84cd4a91003", null ],
+      [ "GetTableType", "struct_pde_table.html#ab09b77683f7a561bac4b30cee2ae895c", null ]
+    ] ],
+    [ "PdeTextRun", "struct_pde_text_run.html", [
+      [ "GetFirstCharIndex", "struct_pde_text_run.html#a71e0a64040bbd32e23594a3effc99bec", null ],
+      [ "GetLastCharIndex", "struct_pde_text_run.html#aeba69525d9c9378a219c2f952a8de43f", null ],
+      [ "GetTextObject", "struct_pde_text_run.html#a6d9b31d647ca9e2bccea1115fff3387b", null ]
+    ] ],
+    [ "PdeWord", "struct_pde_word.html", [
+      [ "GetBackground", "struct_pde_word.html#af97e586a490af7bf93abc06a88a708fc", null ],
+      [ "GetCharBBox", "struct_pde_word.html#a6d6bb4d46890cd2d3afc273298938deb", null ],
+      [ "GetCharCode", "struct_pde_word.html#a26ad53e62003025651b4a6f5b27ebb65", null ],
+      [ "GetCharStateFlags", "struct_pde_word.html#a33a067f8b2c5400d124ef06012919e39", null ],
+      [ "GetCharText", "struct_pde_word.html#a146c9585e2325e263e57c3a9a81a6602", null ],
+      [ "GetCharTextState", "struct_pde_word.html#aa3ba650827ff7923728fa18ba025936e", null ],
+      [ "GetNumChars", "struct_pde_word.html#a83599f8661a44b324ad57c0049d80213", null ],
+      [ "GetNumTextRuns", "struct_pde_word.html#a0e96dc3d39090f0834d3fcf2c643075c", null ],
+      [ "GetOrigin", "struct_pde_word.html#ad3a1fbe50dd7c38802a54a7d54fecd42", null ],
+      [ "GetText", "struct_pde_word.html#aa8b3ea29363b8b8e4da6e6bd1b3ab206", null ],
+      [ "GetTextRun", "struct_pde_word.html#a691f15182df833ee2ea101afc8b01455", null ],
+      [ "GetTextState", "struct_pde_word.html#a8249c3fdf0d030d9b120216ba6b49cad", null ],
+      [ "GetWordFlags", "struct_pde_word.html#a799827c08d965762b4d8dea79cc2c7cd", null ],
+      [ "HasTextState", "struct_pde_word.html#a9a3f6caf9fc6c2186b2f2a602e36e964", null ]
+    ] ],
+    [ "PdeTextLine", "struct_pde_text_line.html", [
+      [ "GetNumWords", "struct_pde_text_line.html#a9b65553fd4f6b494e4318e7d6c8bc6c8", null ],
+      [ "GetText", "struct_pde_text_line.html#abade6b51a09a40957819cf80ae6b2f3a", null ],
+      [ "GetTextLineFlags", "struct_pde_text_line.html#a214d44a4a8192baf06a9d4a8250c25a2", null ],
+      [ "GetTextState", "struct_pde_text_line.html#a136a3588230856cd087947aab34948ac", null ],
+      [ "GetWord", "struct_pde_text_line.html#ae7874c49db5ec67f4e6cde576aa0dbd9", null ],
+      [ "HasTextState", "struct_pde_text_line.html#ae37b3088965683e69c3ba5fd0f86a9b7", null ]
+    ] ],
+    [ "PdeText", "struct_pde_text.html", [
+      [ "GetIndent", "struct_pde_text.html#a91249e2a52cd2777ad769c1b195bc3f7", null ],
+      [ "GetLabelLevel", "struct_pde_text.html#aaff189458efae5a9c605f05984fcca2b", null ],
+      [ "GetLineSpacing", "struct_pde_text.html#a7f98397a239c42d4c6c88a53de4a9233", null ],
+      [ "GetNumTextLines", "struct_pde_text.html#a7c9cf4acbe54bfc7b90d19b755e72026", null ],
+      [ "GetNumWords", "struct_pde_text.html#a05393f024a9abf64c9baaabb0677c6e5", null ],
+      [ "GetText", "struct_pde_text.html#afa2f02f4979cc1bfc4d54cde4323d51a", null ],
+      [ "GetTextFlags", "struct_pde_text.html#a5c8dc70933e65ab900fe043133683f9b", null ],
+      [ "GetTextLine", "struct_pde_text.html#aac1088c094b2dc85ff049a497a72e395", null ],
+      [ "GetTextState", "struct_pde_text.html#a9650096f97c8ca75530112d0d4edc018", null ],
+      [ "GetTextStyle", "struct_pde_text.html#a92f7fd79b25ba6ed768be88fe92f89d7", null ],
+      [ "GetWord", "struct_pde_text.html#afa11ec01b285e86dd114dfb3d1907cbc", null ],
+      [ "HasTextState", "struct_pde_text.html#aad84d19217cdc1a926530f299a78b0f0", null ],
+      [ "SetLabelLevel", "struct_pde_text.html#a98aab8e83047012289e458e74ef10be2", null ]
+    ] ],
+    [ "PdfColorSpace", "struct_pdf_color_space.html", [
+      [ "CreateColor", "struct_pdf_color_space.html#a4342fc41031cb090d6b51239a659f535", null ],
+      [ "GetFamilyType", "struct_pdf_color_space.html#aefdbc589762e0502e544733c6e70d453", null ],
+      [ "GetName", "struct_pdf_color_space.html#ad3f2089a8b6e87d97d5a04883a55efc0", null ],
+      [ "GetNumComps", "struct_pdf_color_space.html#a0d73e262b1128d17863f9fbbeece9c62", null ]
+    ] ],
+    [ "PdfColor", "struct_pdf_color.html", [
+      [ "Destroy", "struct_pdf_color.html#ac48f7be6857d66786ce8e4c9b35e7dc7", null ],
+      [ "GetCMYK", "struct_pdf_color.html#ac8d7e0a542713734d4ae5feaaad89224", null ],
+      [ "GetColorSpace", "struct_pdf_color.html#a4205808ee8526b9cc82e81f99f08f712", null ],
+      [ "GetGrayscale", "struct_pdf_color.html#ad24c3d0ec407c8ca1ae6767ff5649ed1", null ],
+      [ "GetRGB", "struct_pdf_color.html#a1488829a441a3655e6c6597a6122467e", null ],
+      [ "GetValue", "struct_pdf_color.html#a53de4fae3944340bf299d2e952038824", null ],
+      [ "SetColorSpace", "struct_pdf_color.html#abf1086400ad25bea9127a4ba79ab1b4c", null ],
+      [ "SetValue", "struct_pdf_color.html#aaa82a9d3747d53cf7a54da13e18a0ced", null ]
+    ] ],
+    [ "PdePageMap", "struct_pde_page_map.html", [
+      [ "AddTags", "struct_pde_page_map.html#ad79747cbfdcad0c2d52ef58faf266fb1", null ],
+      [ "CreateElement", "struct_pde_page_map.html#a23c0b8d31c40731db9a46ebee9c24090", null ],
+      [ "CreateElements", "struct_pde_page_map.html#a8a36ccac5bc3a5a3f7ebc2df6a25faaa", null ],
+      [ "GetArtifact", "struct_pde_page_map.html#ae21cc70c652692946c383b05f8da21e8", null ],
+      [ "GetBBox", "struct_pde_page_map.html#acc7917f79fc17a2bc35c0e5798ab065e", null ],
+      [ "GetElement", "struct_pde_page_map.html#a074e532585810ca104f159c478c7b371", null ],
+      [ "GetNumArtifacts", "struct_pde_page_map.html#a9c463043c34f8b88584497c741539e88", null ],
+      [ "GetPage", "struct_pde_page_map.html#a9384371f40e761ff0ee436b87c578976", null ],
+      [ "GetWhitespace", "struct_pde_page_map.html#add6d6a726118281d3df54c7cb9ab9108", null ],
+      [ "Release", "struct_pde_page_map.html#ac361d195b0b56716560ab410d39b09be", null ],
+      [ "RemoveElements", "struct_pde_page_map.html#aa38ec8903165fe93da124a8ca75e6cda", null ]
+    ] ]
+];