+++ docs/en/6.13.0/html/search/classes_1.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['ocrtesseract_1330',['OcrTesseract',['../struct_ocr_tesseract.html',1,'']]]
+];