+++ docs/en/6.13.0/html/search/classes_3.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['tesseractdoc_1420',['TesseractDoc',['../struct_tesseract_doc.html',1,'']]]
+];