+++ docs/en/6.13.0/html/search/functions_12.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['unregisterevent_1886',['UnregisterEvent',['../struct_pdfix.html#af47d0c346196335677504c70e74a89b3',1,'Pdfix']]],
+  ['update_1887',['Update',['../struct_pdf_doc_template.html#a862c7c08c7b4bfd0cfdbceee9782e50f',1,'PdfDocTemplate::Update()'],['../struct_ps_standard_authorization.html#aa084066d52f89a79c16d908d02952a47',1,'PsStandardAuthorization::Update()']]],
+  ['updateparenttree_1888',['UpdateParentTree',['../struct_pds_struct_tree.html#a944cbca3c10f1532c213690c3268ee49',1,'PdsStructTree']]]
+];