+++ docs/en/6.13.0/html/search/functions_3.js
@@ -1,0 +1,8 @@
+var searchData=
+[
+  ['deactivate_1494',['Deactivate',['../struct_ps_standard_authorization.html#a0513ab817043f85bbb3c8122c5e76882',1,'PsStandardAuthorization']]],
+  ['deletepages_1495',['DeletePages',['../struct_pdf_doc.html#a21fd1976a10f14a11b877a5649c1d8f3',1,'PdfDoc']]],
+  ['destroy_1496',['Destroy',['../struct_pdf_color.html#ac48f7be6857d66786ce8e4c9b35e7dc7',1,'PdfColor::Destroy()'],['../struct_pdf_security_handler.html#a060cabd72ddcb56bccf8387005286c65',1,'PdfSecurityHandler::Destroy()'],['../struct_pdf_base_dig_sig.html#a2bd43e9a5c9fd84035997d05e4100318',1,'PdfBaseDigSig::Destroy()'],['../struct_ps_regex.html#a1fa82605c116562e9a5051381f70f584',1,'PsRegex::Destroy()'],['../struct_ps_stream.html#a0ab741a3e9e91515d92bc5410ce409a5',1,'PsStream::Destroy()'],['../struct_ps_command.html#a70b7d4f394e503f498a085a57e41fc76',1,'PsCommand::Destroy()'],['../struct_ps_image.html#a676bdafb5a4e0841c9a40c1c9899c536',1,'PsImage::Destroy()'],['../struct_ps_sys_font.html#a227525f2ef6203bed5747e9eb94e65ae',1,'PsSysFont::Destroy()'],['../struct_pdfix.html#a511cb1ecf1fdd4fd8bf49b0a79616e9a',1,'Pdfix::Destroy()'],['../struct_pdfix_plugin.html#a391399f0c4888064501e492f80f1ca0c',1,'PdfixPlugin::Destroy()']]],
+  ['doesmap_1497',['DoesMap',['../struct_pds_role_map.html#a3f8355bcbccb5b8ad8cb10a57d335f62',1,'PdsRoleMap']]],
+  ['drawcontent_1498',['DrawContent',['../struct_pdf_page.html#a9375990421af3049c5399398ee2ea07b',1,'PdfPage']]]
+];