+++ docs/en/6.13.0/html/search/functions_c.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['notifydidchange_1767',['NotifyDidChange',['../struct_pdf_annot.html#a602a1f832d4b7b696410061f168b208d',1,'PdfAnnot::NotifyDidChange()'],['../struct_pdf_form_field.html#a102425a2851532288238a8b88a097c18',1,'PdfFormField::NotifyDidChange()']]],
+  ['notifywillchange_1768',['NotifyWillChange',['../struct_pdf_annot.html#a8cfc1208c6405b14f370b81f5e3e60d9',1,'PdfAnnot::NotifyWillChange()'],['../struct_pdf_form_field.html#af491efbc4c15bd5d03fa67436a18740f',1,'PdfFormField::NotifyWillChange()']]]
+];