+++ docs/en/6.13.0/html/search/functions_e.js
@@ -1,0 +1,15 @@
+var searchData=
+[
+  ['pointinannot_1774',['PointInAnnot',['../struct_pdf_annot.html#a4b16a4c466eef5125854e90bcba42e1a',1,'PdfAnnot']]],
+  ['pointtodevice_1775',['PointToDevice',['../struct_pdf_page_view.html#a1b09261f1ca9be98fe08eca0d8c88fee',1,'PdfPageView']]],
+  ['pointtopage_1776',['PointToPage',['../struct_pdf_page_view.html#afc73415a0eb82230dcab235b1bfd043e',1,'PdfPageView']]],
+  ['put_1777',['Put',['../struct_pds_array.html#ab39831442a3718657a04cea3ab06ab8d',1,'PdsArray::Put()'],['../struct_pds_dictionary.html#abdb7e0652b0fe9c70e1d10448fbf4661',1,'PdsDictionary::Put()']]],
+  ['putarray_1778',['PutArray',['../struct_pds_dictionary.html#a457f52f6a0d2fdc634d69563d53699da',1,'PdsDictionary']]],
+  ['putbool_1779',['PutBool',['../struct_pds_dictionary.html#a68200fd6f25492df740e825415fad6d3',1,'PdsDictionary']]],
+  ['putdict_1780',['PutDict',['../struct_pds_dictionary.html#a3f492f526116e48735405fda56c51d08',1,'PdsDictionary']]],
+  ['putmatrix_1781',['PutMatrix',['../struct_pds_dictionary.html#a29f4c56164f85440c868827a32d97308',1,'PdsDictionary']]],
+  ['putname_1782',['PutName',['../struct_pds_array.html#a93e94babaf6f419faa84b7effa23724c',1,'PdsArray::PutName()'],['../struct_pds_dictionary.html#a4468c5f90703e4c9365ebbe61485368a',1,'PdsDictionary::PutName()']]],
+  ['putnumber_1783',['PutNumber',['../struct_pds_array.html#af88da39b9ecb8fa94878c642703cf6a8',1,'PdsArray::PutNumber()'],['../struct_pds_dictionary.html#a7c254f22d9fe667316f90b1da6bbd0f5',1,'PdsDictionary::PutNumber()']]],
+  ['putrect_1784',['PutRect',['../struct_pds_dictionary.html#a4690b1bb40264ac275cde3c8ee6f818a',1,'PdsDictionary']]],
+  ['putstring_1785',['PutString',['../struct_pds_array.html#afed481450fb3196c3d1fd8045e1e5263',1,'PdsArray::PutString()'],['../struct_pds_dictionary.html#aac38660e8a1f41bc99f6825bd20105e0',1,'PdsDictionary::PutString()']]]
+];