+++ docs/en/6.13.0/html/search/groups_2.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['ocrtesseract_20enumerations_2597',['OcrTesseract Enumerations',['../group___enumerations___p_d_fix___ocr___tesseract.html',1,'']]],
+  ['ocrtesseract_20sdk_2598',['OcrTesseract SDK',['../group___ocr_tesseract___p_d_fix___ocr___tesseract.html',1,'']]],
+  ['ocrtesseract_20structures_2599',['OcrTesseract Structures',['../group___structures___p_d_fix___ocr___tesseract.html',1,'']]]
+];