+++ docs/en/6.13.0/html/search/pages_0.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['pdfix_20sdk_20documentation_2610',['PDFix SDK Documentation',['../index.html',1,'']]],
+  ['pdfix_20app_2611',['PDFix App',['../pdfixcli.html',1,'']]]
+];