+++ docs/en/6.13.0/html/search/variables_c.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['opacity_1941',['opacity',['../struct___pdf_annot_appearance.html#adbdf78e0e0e502350b9667cc72c74aee',1,'_PdfAnnotAppearance']]]
+];