+++ docs/en/6.13.0/html/search/variables_d.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['permissions_1942',['permissions',['../struct___pdf_standard_security_params.html#a02d6fa0f26779e43695aaf86070d52ab',1,'_PdfStandardSecurityParams']]]
+];