+++ docs/en/6.13.0/html/struct___pdf_c_m_y_k.js
@@ -1,0 +1,10 @@
+var struct___pdf_c_m_y_k =
+[
+    [ "_PdfCMYK", "struct___pdf_c_m_y_k.html#ab6d33e79afcb3ad68c422c13465b6881", null ],
+    [ "operator<", "struct___pdf_c_m_y_k.html#acec826ea7144660f2fc94b2eb941e41e", null ],
+    [ "operator==", "struct___pdf_c_m_y_k.html#a41e6d9f72e3f3c0369082957026be747", null ],
+    [ "c", "struct___pdf_c_m_y_k.html#ae0d7fb63f3f8ee8e9dafec2b1c172553", null ],
+    [ "k", "struct___pdf_c_m_y_k.html#a502ef6c5a89f3aa5cf5eddeb755094af", null ],
+    [ "m", "struct___pdf_c_m_y_k.html#ae36c451da5e76738af0d7249401f9e3e", null ],
+    [ "y", "struct___pdf_c_m_y_k.html#acaec2cfacc88d2706a0e6c14ff7dba61", null ]
+];