+++ docs/en/6.13.0/html/struct___pdf_dev_point.js
@@ -1,0 +1,6 @@
+var struct___pdf_dev_point =
+[
+    [ "_PdfDevPoint", "struct___pdf_dev_point.html#afd50b10697a2164d1b3db027fd330afb", null ],
+    [ "x", "struct___pdf_dev_point.html#ad21f7fa54179220e39ee06e97f00e3c1", null ],
+    [ "y", "struct___pdf_dev_point.html#a41c85bb5296685ebbedf668a5e1a0fdc", null ]
+];