+++ docs/en/6.13.0/html/struct_pdf_action.js
@@ -1,0 +1,8 @@
+var struct_pdf_action =
+[
+    [ "GetDestFile", "struct_pdf_action.html#ab5e5004fa2e51cc8d199e7278bfd458d", null ],
+    [ "GetJavaScript", "struct_pdf_action.html#ad60d7ba0524bb24e504b714aeb1cfa55", null ],
+    [ "GetSubtype", "struct_pdf_action.html#a772509158dbc224df879e91aa44897cb", null ],
+    [ "GetViewDestination", "struct_pdf_action.html#abd48115d5a719c8b87b6b79f70f45562", null ],
+    [ "Release", "struct_pdf_action.html#a65046e66e52f566f7134f39cf51e8410", null ]
+];