+++ docs/en/6.13.0/html/struct_pdf_bookmark.js
@@ -1,0 +1,20 @@
+var struct_pdf_bookmark =
+[
+    [ "AddNewChild", "struct_pdf_bookmark.html#a8b0dc672e0afed978b198e9b44fd74de", null ],
+    [ "GetAction", "struct_pdf_bookmark.html#a0af8b9a9391aaf001e22d4c9fc3f4494", null ],
+    [ "GetAppearance", "struct_pdf_bookmark.html#a5494ca04ed4061d10d04ddea849069f2", null ],
+    [ "GetChild", "struct_pdf_bookmark.html#a972a1ef67fa50cf3042761a5ad5ce4bd", null ],
+    [ "GetNext", "struct_pdf_bookmark.html#a4ad7ab7cb9fe459d2147b993894bf252", null ],
+    [ "GetNumChildren", "struct_pdf_bookmark.html#a0ce73f73eff9da7770a0a03925389adf", null ],
+    [ "GetObject", "struct_pdf_bookmark.html#a5252227598de64deaf41c80aa2118fea", null ],
+    [ "GetParent", "struct_pdf_bookmark.html#ad77a188c0e5a79ecb2cd26164eb5e3cb", null ],
+    [ "GetPrev", "struct_pdf_bookmark.html#ada18236d6e0653f9afdf35ca86fd5216", null ],
+    [ "GetTitle", "struct_pdf_bookmark.html#acc1510f122ac4394ee3236082a8a813e", null ],
+    [ "IsOpen", "struct_pdf_bookmark.html#ad7b002178bafe1764ef7c56ab5783b38", null ],
+    [ "IsValid", "struct_pdf_bookmark.html#ad8c4fdcfbf2f71a5c71fedc32adefd29", null ],
+    [ "Remove", "struct_pdf_bookmark.html#ac0a6882411d97f309db5a55bf7e94dfe", null ],
+    [ "SetAction", "struct_pdf_bookmark.html#a20df7d356f69270b22a0f09343502431", null ],
+    [ "SetAppearance", "struct_pdf_bookmark.html#a802da8c359c7b9499d21233be4d7feee", null ],
+    [ "SetOpen", "struct_pdf_bookmark.html#af453fae7bf7e5cdffd09bb1285c675fe", null ],
+    [ "SetTitle", "struct_pdf_bookmark.html#acf394787aeedbb5a2d728313729a3ce6", null ]
+];