+++ docs/en/6.13.0/html/struct_pdf_doc_undo.js
@@ -1,0 +1,9 @@
+var struct_pdf_doc_undo =
+[
+    [ "BeginOperation", "struct_pdf_doc_undo.html#aeed8550cffb0688d67bfc152dede8bd0", null ],
+    [ "EndOperation", "struct_pdf_doc_undo.html#a0095bb55e4dcfa3dba8ee922cd8e4f31", null ],
+    [ "Execute", "struct_pdf_doc_undo.html#a336c400d90dc4345d906b61ad46cfd8b", null ],
+    [ "GetData", "struct_pdf_doc_undo.html#a072f40c1b52d9264b0cceb82c91f5b50", null ],
+    [ "GetNumEntries", "struct_pdf_doc_undo.html#ade788a25d5bfc3d29a42d0706280f87c", null ],
+    [ "GetTitle", "struct_pdf_doc_undo.html#ab06996f32624ba606038f023dc4f2383", null ]
+];