+++ docs/en/6.13.0/html/struct_pds_dictionary.js
@@ -1,0 +1,25 @@
+var struct_pds_dictionary =
+[
+    [ "Get", "struct_pds_dictionary.html#a2d8b648c8712cfe8dfc84c03bd7ac182", null ],
+    [ "GetArray", "struct_pds_dictionary.html#a9cebb3d736406b834368c460e2139007", null ],
+    [ "GetBoolean", "struct_pds_dictionary.html#ab29665202ba4ad328066a2c9787c28e3", null ],
+    [ "GetDictionary", "struct_pds_dictionary.html#a0ea942dcc26da8f58d86f8e792fd9ea0", null ],
+    [ "GetInteger", "struct_pds_dictionary.html#a4a04d59929f92449daa7fe7e76f81219", null ],
+    [ "GetKey", "struct_pds_dictionary.html#a324ae471f8371c27c2b8e3b052ad6df9", null ],
+    [ "GetNumber", "struct_pds_dictionary.html#a886801f943e19e97faa8f0c8e630de61", null ],
+    [ "GetNumKeys", "struct_pds_dictionary.html#acd150baef2d6f81141347d71d4fe4bb9", null ],
+    [ "GetStream", "struct_pds_dictionary.html#a7c90a0554106b96bb66fda9a815a55b6", null ],
+    [ "GetString", "struct_pds_dictionary.html#aa19659181503da64a78c4fa30c6fe669", null ],
+    [ "GetText", "struct_pds_dictionary.html#a0ed655a04211dc831369315f72f09fe8", null ],
+    [ "Known", "struct_pds_dictionary.html#ade2629fdf828bce3529991bd27ca48db", null ],
+    [ "Put", "struct_pds_dictionary.html#abdb7e0652b0fe9c70e1d10448fbf4661", null ],
+    [ "PutArray", "struct_pds_dictionary.html#a457f52f6a0d2fdc634d69563d53699da", null ],
+    [ "PutBool", "struct_pds_dictionary.html#a68200fd6f25492df740e825415fad6d3", null ],
+    [ "PutDict", "struct_pds_dictionary.html#a3f492f526116e48735405fda56c51d08", null ],
+    [ "PutMatrix", "struct_pds_dictionary.html#a29f4c56164f85440c868827a32d97308", null ],
+    [ "PutName", "struct_pds_dictionary.html#a4468c5f90703e4c9365ebbe61485368a", null ],
+    [ "PutNumber", "struct_pds_dictionary.html#a7c254f22d9fe667316f90b1da6bbd0f5", null ],
+    [ "PutRect", "struct_pds_dictionary.html#a4690b1bb40264ac275cde3c8ee6f818a", null ],
+    [ "PutString", "struct_pds_dictionary.html#aac38660e8a1f41bc99f6825bd20105e0", null ],
+    [ "RemoveKey", "struct_pds_dictionary.html#ab607e7442f8eb77b01d91af4c6221a59", null ]
+];